+++ src/html.js
@@ -39,7 +39,6 @@ function cleanDocument (document) {
 
   return document;
 }
-cleanDocument({});
 
 /**
  * Replaces all instances of a class name in class attributes in the DOM
@@ -73,7 +72,6 @@ function replaceSemanticClassWithAtomizedClasses (node, classToReplace, newClass
   if (node.attrs) {
     node.attrs.forEach(function (attribute) {
       if (attribute.name === 'class') {
-        // console.log(JSON.stringify(cleanDocument(node), null, 2));
         // 'cool cat nice wow' => ['cool','cat','nice','wow']
         let classes = attribute.value.split(' ');
         if (classes.includes(classToReplace)) {
@@ -115,7 +113,7 @@ const html = function (options, input, classMap) {
   classMap = classMap || {};
 
   // String => Object
-  const document = parse5.parse(input);
+  const document = cleanDocument(parse5.parse(input));
 
   Object.keys(classMap).forEach(function (semanticClass) {
     let atomizedClasses = classMap[semanticClass];
@@ -127,8 +125,6 @@ const html = function (options, input, classMap) {
     }
     replaceSemanticClassWithAtomizedClasses(document, semanticClass, atomizedClasses);
   });
-
-  // console.log(JSON.stringify(cleanDocument(document), null, 2));
 
   // Object => string
   let markup = parse5.serialize(document);
